--- backend/src/model/NewsletterHeader.ts
@@ -1,3 +1,0 @@
-export class NewsletterHeader {
-  [key: string]: string
-}
