--- backend/src/model/NewsletterParams.ts
@@ -1,4 +1,0 @@
-export class NewsletterParams {
-  parameter: string
-  subject: string
-}
