--- backend/src/model/NewsletterUser.ts
@@ -1,4 +1,0 @@
-export class NewsletterUser {
-  name: string
-  email: string
-}
